+++ Engine/Memory/DynamicAllocator.h
@@ -22,7 +22,7 @@ public:
 	 * @param total_size The total size in bytes the allocator should hold. Note this size does not include the size of the internal state.
 	 * @return True on success.
 	 */
-	DAPI bool Create(unsigned long long total_size);
+	DAPI bool Create(size_t total_size);
 
 	/**
 	 * @brief Destroys the allocator.
@@ -37,7 +37,7 @@ public:
 	 * @param size The size in bytes to be allocated.
 	 * @return The allocated block of memory unless this operation fails, then nullptr.
 	 */
-	DAPI void* Allocate(unsigned long long size);
+	DAPI void* Allocate(size_t size);
 
 	/**
 	 * @brief Free the given block of memory.
@@ -45,7 +45,7 @@ public:
 	 * @param block The block to be freed. Must have been allocated by the allocator.
 	 * @param size The size in bytes to be allocated.
 	 */
-	DAPI bool Free(void* block, unsigned long long size);
+	DAPI bool Free(void* block, size_t size);
 
 	/**
 	 * @brief Obtains the amount of free space left in the allocator.
@@ -55,7 +55,7 @@ public:
 	DAPI unsigned long long GetFreeSpace();
 
 private:
-	unsigned long long TotalSize;
+	size_t TotalSize;
 	Freelist List;
 	void* MemoryBlock;
 };
